requirements.txt
streamlit
numpy
tensorflow-cpu
Pillow
stardist
scipy
pandas
matplotlib
skimage
scikit-learn